PortStatus と Statusを commonMain に移動
