+++ app/src/main/java/com/ikmr/banbara23/yaeyama_liner_checker/api/ApiClient.kt
@@ -2,9 +2,9 @@ package com.ikmr.banbara23.yaeyama_liner_checker.api
 
 import com.google.firebase.database.DatabaseReference
 import com.google.firebase.database.FirebaseDatabase
+import com.ikemura.shared.model.statusdetail.PortStatus
 import com.ikmr.banbara23.yaeyama_liner_checker.model.Company
 import com.ikmr.banbara23.yaeyama_liner_checker.model.DetailLinerInfo
-import com.ikmr.banbara23.yaeyama_liner_checker.model.PortStatus
 import com.ikmr.banbara23.yaeyama_liner_checker.model.StatusDetailRoot
 import com.ikmr.banbara23.yaeyama_liner_checker.model.time_table.TimeTable
 import durdinapps.rxfirebase2.RxFirebaseDatabase
