+++ app/src/main/java/com/ikmr/banbara23/yaeyama_liner_checker/model/time_table/TimeTable.kt
@@ -1,10 +1,6 @@
 package com.ikmr.banbara23.yaeyama_liner_checker.model.time_table
 
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
 data class TimeTable(
     val header: Header = Header(),
     val row: List<Row> = listOf()
-) : Parcelable
+)
