+++ app/src/main/java/com/ikmr/banbara23/yaeyama_liner_checker/ui/portlist/PortListAdapter.kt
@@ -3,16 +3,12 @@ package com.ikmr.banbara23.yaeyama_liner_checker.ui.portlist
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.RecyclerView
+import com.ikemura.shared.model.statusdetail.PortStatus
 import com.ikmr.banbara23.yaeyama_liner_checker.R
-import com.ikmr.banbara23.yaeyama_liner_checker.model.PortStatus
-import kotlinx.android.synthetic.main.port_list_item.view.description
-import kotlinx.android.synthetic.main.port_list_item.view.port_name
-import kotlinx.android.synthetic.main.port_list_item.view.status
 
 class PortListAdapter(
     lifecycleOwner: LifecycleOwner,
@@ -36,9 +32,9 @@ class PortListAdapter(
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         val item = items[position]
-        holder.portName.text = item.portName
-        holder.description.text = item.comment
-        holder.status.text = item.status.text
+        // holder.portName.text = item.portName
+        // holder.description.text = item.comment
+        // holder.status.text = item.status.text
 
         with(holder.view) {
             tag = item
@@ -56,12 +52,9 @@ class PortListAdapter(
     override fun getItemCount(): Int = items.size
 
     inner class ViewHolder(val view: View) : RecyclerView.ViewHolder(view) {
-        val portName: TextView = view.port_name
-        val description: TextView = view.description
-        val status: TextView = view.status
+        // val portName: TextView = view.port_name
+        // val description: TextView = view.description
+        // val status: TextView = view.status
 
-        override fun toString(): String {
-            return super.toString() + " '" + portName.text + "'" + " '" + description.text + "'" + " '" + status.text + "'"
-        }
     }
 }
