+++ app/src/main/java/com/ikmr/banbara23/yaeyama_liner_checker/ui/portlist/PortListFragment.kt
@@ -9,8 +9,8 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.ikemura.shared.model.statusdetail.PortStatus
 import com.ikmr.banbara23.yaeyama_liner_checker.R
-import com.ikmr.banbara23.yaeyama_liner_checker.model.PortStatus
 
 class PortListFragment : Fragment() {
