+++ app/src/main/java/com/ikmr/banbara23/yaeyama_liner_checker/ui/portlist/PortListViewModel.kt
@@ -2,7 +2,7 @@ package com.ikmr.banbara23.yaeyama_liner_checker.ui.portlist
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel;
-import com.ikmr.banbara23.yaeyama_liner_checker.model.PortStatus
+import com.ikemura.shared.model.statusdetail.PortStatus
 
 class PortListViewModel : ViewModel() {
     val items = MutableLiveData<List<PortStatus>>()
