+++ app/src/main/res/layout/view_status_detail_top.xml
@@ -15,7 +15,7 @@
 
         <variable
             name="portStatus"
-            type="com.ikmr.banbara23.yaeyama_liner_checker.model.PortStatus"
+            type="com.ikemura.shared.model.statusdetail.PortStatus"
             />
     </data>
