+++ shared/src/commonMain/kotlin/com/ikemura/shared/model/statusdetail/PortStatus.kt
@@ -1,0 +1,8 @@
+package com.ikemura.shared.model.statusdetail
+
+data class PortStatus(
+    val comment: String = "",
+    val portCode: String = "",
+    val portName: String = "",
+    val status: Status = Status()
+)
