+++ shared/src/commonMain/kotlin/com/ikemura/shared/model/statusdetail/Status.kt
@@ -1,0 +1,6 @@
+package com.ikemura.shared.model.statusdetail
+
+data class Status(
+    val code: String = "",
+    val text: String = ""
+)
